Implemented lambda/rocco@b39f8216d630509d2f77ebf48d4e81309f4d7ea6 Allow user to set encoding of files, transcoding to UTF-8
Ruby 1.9 is much more strict about encodings; instead of just reading bytes
into the string and hoping you know how to interpret it, it actually keeps
track of what encoding it expects the file and the string to be and
transcodes between them. This means that we can no longer just read in the
ISO-8859-1 as binary, and kind of hope that everything later on ignores it
or knows how to cope. Instead, we need to explicitly choose to read the
file as ISO-8859-1, and explicitly choose the encoding that we want the
string to be in.  We use UTF-8, because that's what Pygments and at least
some of the Markdown libraries are expecting.

Fixes: https://github.com/rtomayko/rocco/issues/73 (fix rtomayko/rocco#73)