Display output even if there are no comments
Rocco should display the code even if there are no comments.  When there are
no comments, it produces a single section with an empty docs string.
However, to pass it through Markdown, it joins the sections with a
delimiter, munges that, and split the result.  Joining a list continaing the
empty string and splitting it again produces an empty list, not a list
containting the empty string:

 > [""].join("delim").split("delim")
 => []

In Ruby 1.9, this causes us lose the associated code when when try to zip
these lists back together again:

 > [].zip(["something"])
 => []

To fix this, just make sure `docs_html` contains at least a single empty
string if it winds up being empty.

Fixes: https://github.com/rtomayko/rocco/issues/71 (fix rtomayko/rocco#71)